Split txpool validator tests
